ChannelPolicy.checkEvent(policy, event)
