+++ packages/pico-engine-core/src/ChannelPolicy.js
@@ -56,6 +56,38 @@ var clean = function(policy_orig){
     return policy;
 };
 
+
+var doesMatchEvent = function(events, event){
+
+    if(events === "ALL"){
+        return true;
+    }
+
+    return _.find(events, function(s){
+        if(_.has(s, "domain") && s.domain !== event.domain){
+            return false;
+        }
+        if(_.has(s, "type") && s.type !== event.type){
+            return false;
+        }
+        return true;
+    });
+};
+
+
+var checkEvent = function(policy, event){
+    if( ! policy || ! policy.events){
+        //TODO remove this
+        return true;
+    }
+    if(policy.events.type === "whitelist"){
+        return doesMatchEvent(policy.events.events, event);
+    }
+    return ! doesMatchEvent(policy.events.events, event);
+};
+
+
 module.exports = {
     clean: clean,
+    checkEvent: checkEvent,
 };
